move more networking utils
