+++ pkg/network/subnets.go
@@ -1,22 +1,22 @@
 // SPDX-License-Identifier: APACHE-2.0
 
-package main
+package network
 
 import (
 	"net"
 	"strings"
 )
 
-// splitCIDRs given a comma separated list with CIDRS it returns 2 slice of strings per IP family
-func splitCIDRs(cidrs string) ([]string, []string) {
+// SplitCIDRs given a comma separated list with CIDRS it returns 2 slice of strings per IP family
+func SplitCIDRs(cidrs string) ([]string, []string) {
 	if cidrs == "" {
 		return nil, nil
 	}
 	subnets := strings.Split(cidrs, ",")
-	return splitCIDRslice(subnets)
+	return SplitCIDRslice(subnets)
 }
 
-func splitCIDRslice(cidrs []string) ([]string, []string) {
+func SplitCIDRslice(cidrs []string) ([]string, []string) {
 	var v4subnets, v6subnets []string
 	for _, subnet := range cidrs {
 		if isIPv6CIDRString(subnet) {
